+++ src/main/resources/META-INF/changelog.md
@@ -9,6 +9,7 @@
 - Copy-as-image (Shift+Shortcut+C) has been replaced with a copy format quick-picker, so you don't have to go into preferences and change your format.
 - Paste-as-format (Shift+Shortcut+V) has been added, much like the above.
 - You can now copy deck images as JPEGs, if you're tired of Discord complaining about your image upload sizes. (They're temporary files that are deleted when the program is closed.)
+- Added "Copy Image" option to most card context menus, to quickly copy a card image to your clipboard.
 
 ### Improvements